backend/requirements.txt
flask
gunicorn
psycopg2-binary
llama-cpp-python